+++ src/app/(home)/(posts)/_components/conference/ForumPosts.tsx
@@ -1,36 +1,13 @@
 'use client';
 
-import { UUID } from 'crypto';
-import { useEffect, useState } from 'react';
-
-type Post = {
-  id: UUID;
-  title: string;
-  content: string;
-  created_at: string;
-};
-
-const fetchForumPosts = async (): Promise<Post[]> => {
-  try {
-    const response = await fetch('/api/posts/conference');
-    const data = await response.json();
-    return data.data;
-  } catch (error) {
-    console.error('Error Fetching Forum Posts', error);
-    return [];
-  }
-};
+import useFetchForumPosts from '@/hooks/conference/useFetchForumPosts';
 
 const ForumPosts = () => {
-  const [posts, setPosts] = useState<Post[]>([]);
+  const { data: posts = [], error, isPending, isStale } = useFetchForumPosts();
 
-  useEffect(() => {
-    const getPosts = async () => {
-      const fetchedPosts = await fetchForumPosts();
-      setPosts(fetchedPosts);
-    };
-    getPosts();
-  }, []);
+  if (isStale && isPending) return <div>Loading...</div>;
+
+  if (error) return <div>Error Fetching Posts</div>;
 
   return (
     <div>
