+++ src/types/posts/forumPosts.ts
@@ -1,0 +1,6 @@
+export type Post = {
+  id: string;
+  title: string;
+  content: string;
+  created_at: string;
+};
